refactor(cli): extract lifecycle handlers — phase 2 redo on post-#1049 base (card df8f6723, originally 5aa9e780)
## Why

Original phase 2 (#1046) closed due to rebase conflict against post-#1049 rust-rewrite. Same pattern as phase 1 redo (#1063): hard-reset to current tip, re-extract cleanly.

## What this PR lands

The card-state-machine handlers move to a sibling module `work_commands_lifecycle.rs`:

  - run_create, run_seed
  - run_claim, run_release, resolve_my_active_claim
  - run_heartbeat, run_update, run_state, run_close
  - cli_can_set_state_directly (state guard)
  - close_transition_allowed_from (state guard)
  - refusal_message (per-state actionable text)

## File sizes

  work_commands.rs:           1713 → 1356 LOC (-357)
  work_commands_lifecycle.rs:    0 →  ~370 LOC (new)

## Cross-module callees

  - spawn_claim_worktree, parse_work_card_id, parse_claim_id,
    parse_optional_lane_id, now_ms, auto_spawn_review_card,
    open_pr_and_link — all stay in work_commands.rs (not yet
    extracted at this base), reached via crate::work_commands::
    with pub(crate) visibility.

  - When phase 3 (git module) + phase 5 (review module) land,
    spawn_claim_worktree + auto_spawn_review_card move to their
    respective modules; this module's callsites pick that up
    automatically because they go through crate::work_commands::
    which can re-export.

## Public surface unchanged

  pub use crate::work_commands_lifecycle::{
      run_claim, run_close, run_create, run_heartbeat, run_release,
      run_seed, run_state, run_update,
  };

main.rs dispatcher sees the same `work_commands::run_*` paths.

## Tests + lints

  cargo test work_commands:: → 20 pass
  cargo clippy -p airc-cli --bin airc -- -D warnings → clean
  cargo fmt --check → clean

## Anchor

c0bd865c follow-up sequence. Phase 1 (#1063, render) in flight; this PR is phase 2. Phase 3 (git) and Phase 5 (review) closed pending redo.
